Enable mixed-precision training

+++ src/tasks/classification/erm.py
@@ -36,9 +36,9 @@ class ERMTask(ClassificationTask[ERM_X]):
         )
 
     def _log_images(self, writer: SummaryWriter, eval_output: ClassificationEvalOutput[ERM_X], *, prefix: str) -> None:
-        eval_output_x = eval_output.x.detach().cpu()
-        eval_output_y = eval_output.y.detach().cpu()
-        eval_output_y_hat = eval_output.y_hat.detach().cpu()
+        eval_output_x = eval_output.x.detach().cpu().float()
+        eval_output_y = eval_output.y.detach().cpu().float()
+        eval_output_y_hat = eval_output.y_hat.detach().cpu().float()
         batch_size, num_classes = eval_output_y_hat.shape
 
         nrows = batch_size
